Function confirm delete and css styles are added

+++ src/components/products/ProductItem.js
@@ -1,40 +1,63 @@
 import React from "react";
-import { FaTrash as DeleteIcon, FaPen as EditIcon } from "react-icons/fa";
+import EditIcon from "@mui/icons-material/Edit";
+import DeleteIcon from "@mui/icons-material/Delete";
 import { useDispatch } from "react-redux";
 import { deleteAsyncCreator } from "../../redux/slices/productsSlice";
 
 export const ProductItem = ({ productToShow, onEdit }) => {
   const dispatch = useDispatch();
+  const Swal = require("sweetalert2");
 
-  const handleDelete = (_id) => {
-    const action = deleteAsyncCreator(productToShow._id);
-    dispatch(action);
+  const handleDelete = () => {
+    Swal.fire({
+      title: "Wait!",
+      text: "Are you sure you want to delete this Product?",
+      icon: "warning",
+      showCancelButton: true,
+      confirmButtonText: "Yes",
+      cancelButtonText: "No",
+    }).then((result) => {
+      if (result.isConfirmed) {
+        Swal.fire({
+          title: "Succes!",
+          text: "Product delete!",
+          icon: "success",
+        });
+        const action = deleteAsyncCreator(productToShow._id);
+        dispatch(action);
+      }
+    });
   };
 
   return (
-    <tbody>
-      <tr>
-        <td>{productToShow.name}</td>
-        <td>{productToShow.description}</td>
-        <td>{productToShow.price}</td>
-        <td>{productToShow.brand}</td>
-        <td>
-          <div className="Edit_Delete">
-            <div>
-              <EditIcon
-                onClick={() => onEdit(productToShow)}
-                style={{ cursor: "pointer", color: "#944ca8" }}
-              />
-            </div>
-            <div>
-              <DeleteIcon
-                onClick={handleDelete}
-                style={{ cursor: "pointer", color: "#7c3494" }}
-              />
-            </div>
-          </div>
-        </td>
-      </tr>
-    </tbody>
+    <div className="bodyTable">
+      <div
+        style={{
+          width: "90%",
+          display: "flex",
+        }}
+      >
+        <div className="field">{productToShow.name}</div>
+        <div className="field">{productToShow.description}</div>
+        <div className="field">{productToShow.price}</div>
+        <div className="field">{productToShow.brand}</div>
+      </div>
+      <div
+        style={{
+          width: "10%",
+          display: "flex",
+        }}
+      >
+        <EditIcon
+          onClick={() => onEdit(productToShow)}
+          style={{ cursor: "pointer", color: "#944ca8", marginRight: "1rem" }}
+        />
+
+        <DeleteIcon
+          onClick={handleDelete}
+          style={{ cursor: "pointer", color: "#7c3494" }}
+        />
+      </div>
+    </div>
   );
 };
